rebranded new global code
