+++ global.c
@@ -158,3 +158,25 @@ unsigned* sort_uints_by_category(
 {
   GENERIC_SORT_BY_CATEGORY(unsigned, a, width, n, cats, indices, cat_offsets);
 }
+
+#define GENERIC_UNSORT_BY_CATEGORY(T, a, width, n, cats, indices, cat_offsets) \
+  T* out = LOOP_MALLOC(T, n); \
+  for (unsigned i = 0; i < n; ++i) { \
+    unsigned cat = cats[i]; \
+    unsigned idx = indices[i]; \
+    unsigned o = cat_offsets[cat]; \
+    for (unsigned j = 0; j < width; ++j) \
+      out[i * width + j] = a[(o + idx) * width + j]; \
+  } \
+  return out;
+
+unsigned* unsort_uints_by_category(
+    unsigned const* a,
+    unsigned width,
+    unsigned n,
+    unsigned const* cats,
+    unsigned const* indices,
+    unsigned const* cat_offsets)
+{
+  GENERIC_UNSORT_BY_CATEGORY(unsigned, a, width, n, cats, indices, cat_offsets);
+}
