+++ global.h
@@ -20,5 +20,12 @@ unsigned* sort_uints_by_category(
     unsigned const* cats,
     unsigned const* indices,
     unsigned const* cat_offsets);
+unsigned* unsort_uints_by_category(
+    unsigned const* a,
+    unsigned width,
+    unsigned n,
+    unsigned const* cats,
+    unsigned const* indices,
+    unsigned const* cat_offsets);
 
 #endif
