+++ owners_from_global.h
@@ -1,0 +1,10 @@
+#ifndef OWNERS_FROM_GLOBAL_H
+#define OWNERS_FROM_GLOBAL_H
+
+void owners_from_global(
+    unsigned n,
+    unsigned long* global_in,
+    unsigned** p_own_parts,
+    unsigned** p_own_idxs);
+
+#endif
